Build i386 binary

+++ ddist.sh
@@ -2,17 +2,23 @@
 set -o errexit
 set -o nounset
 
+if [[ "$#" != 1 ]]; then
+  echo "Usage: $0 ARCH_SUFFIX"
+  exit 1
+fi
+suffix="$1"
+
 REL_HERE=$(dirname "${BASH_SOURCE}")
 HERE=$(cd "${REL_HERE}"; pwd)
 
-IMG="tini"
+IMG="tini-build-${suffix}"
 SRC="/tini"
 
 # Cleanup the build dir
 rm -f "${HERE}/dist"/*
 
 # Create the build image
-docker build -t "${IMG}" .
+docker build --build-arg "ARCH_SUFFIX=${suffix}" -t "${IMG}" .
 
 # Run test without subreaper support, don't copy build files here
 docker run -it --rm \
@@ -22,7 +28,8 @@ docker run -it --rm \
   -e FORCE_SUBREAPER="${FORCE_SUBREAPER:="1"}" \
   -e GPG_PASSPHRASE="${GPG_PASSPHRASE:=}" \
   -e CC="${CC:=gcc}" \
+  -e CFLAGS="${CFLAGS-}" \
   -e ARCH_NATIVE="${ARCH_NATIVE-1}" \
-  -e ARCH_SUFFIX="${ARCH_SUFFIX-}" \
+  -e ARCH_SUFFIX="${suffix}" \
   -e MINIMAL="${MINIMAL-}" \
   "${IMG}" "${SRC}/ci/run_build.sh"
